Splitted some more code. Added validators for numberUtils & validators.
