+++ src/utils/numberUtils.ts
@@ -1,0 +1,5 @@
+const sum = (a: number, b: number): number => a + b;
+
+const average = (list: number[]): number => list.reduce(sum, 0) / list.length;
+
+export { sum, average };
