+++ src/weboodi.ts
@@ -1,3 +1,4 @@
+import { kurssitietokanta } from "./data/courses";
 import {
   ConvertedCourse,
   CourseArrToObjParams,
@@ -10,6 +11,8 @@ import {
   getLocalStorage,
   setLocalStorage,
 } from "./utils/localStorage";
+import { average, sum } from "./utils/numberUtils";
+import { isArray, isString, isTruthy } from "./utils/validators";
 
 const setDuplikaattiKurssit = setLocalStorage<string[]>("duplikaattiKurssit");
 const setPerusOpinnot = setLocalStorage<string[]>("perusOpinnot");
@@ -33,12 +36,6 @@ interface Paivays {
 
 const findPvm = <T>(list: Array<T & Paivays>, key: string) =>
   list.find((val) => val.pvm === key);
-
-const isTruthy = (v) => v;
-
-const isString = (val: unknown): val is string => typeof val === "string";
-
-const isArray = <T>(val: unknown): val is T[] => Array.isArray(val);
 
 const toLowerCase = (str: string) => str.toLowerCase();
 
@@ -80,159 +77,6 @@ const chartColors = [
   "purple",
 ];
 
-const kurssitietokanta = {
-  tkt: {
-    perusopinnot: [
-      {
-        name: "Johdatus tietojenkäsittelytieteeseen",
-        keys: ["TKT10001", "AYTKT10001", "582102", "A582102"],
-      },
-      {
-        name: "Ohjelmoinnin perusteet",
-        keys: ["TKT10002", "AYTKT10002", "581325", "A581325"],
-      },
-      {
-        name: "Ohjelmoinnin jatkokurssi",
-        keys: ["TKT10003", "AYTKT10003", "582103", "A582103"],
-      },
-      {
-        name: "Tietokantojen perusteet",
-        keys: ["TKT10004", "AYTKT10004", "581328", "A581328"],
-      },
-      {
-        name: "Tietokoneen toiminta",
-        keys: ["TKT10005", "AYTKT10005", "581305", "A581305"],
-      },
-    ],
-    aineopinnot: [
-      {
-        name: "Tietorakenteet ja algoritmit",
-        keys: ["TKT20001", "AYTKT20001", "58131", "A58131"],
-      },
-      {
-        name: "Ohjelmistotekniikan menetelmät",
-        keys: ["TKT20002", "AYTKT20002", "582104", "A582104"],
-      },
-      {
-        name: "Käyttöjärjestelmät",
-        keys: ["TKT20003", "AYTKT20003", "582219", "A582219"],
-      },
-      {
-        name: "Tietoliikenteen perusteet",
-        keys: ["TKT20004", "AYTKT20004", "582202", "A582202"],
-      },
-      {
-        name: "Laskennan mallit",
-        keys: ["TKT20005", "AYTKT20005", "582206", "A582206"],
-      },
-      {
-        name: "Ohjelmistotuotanto",
-        keys: ["TKT20006", "AYTKT20006", "581259", "A581259"],
-      },
-      {
-        name: "Ohjelmistotuotantoprojekti",
-        keys: ["TKT20007", "AYTKT20007", "581260", "A581260"],
-      },
-      {
-        name: "Kandidaatin tutkielma",
-        keys: ["TKT20013", "AYTKT20013", "582204", "A582204"],
-      },
-      {
-        name: "Kypsyysnäyte LuK",
-        keys: ["TKT20014", "AYTKT20014", "50036", "A50036"],
-      },
-    ],
-    mitaNaaOn: [
-      {
-        name: "Johdatus tekoälyyn",
-        keys: ["TKT20008", "AYTKT20008", "582216", "A582216"],
-      },
-      {
-        name: "Tietoturvan perusteet",
-        keys: ["TKT20009", "AYTKT20009", "582215", "A582215"],
-      },
-    ],
-    perusopinnotPre2017: [
-      {
-        name: "Johdatus tietojenkäsittelytieteeseen",
-        keys: ["TKT10001", "AYTKT10001", "582102", "A582102"],
-      },
-      {
-        name: "Ohjelmoinnin perusteet",
-        keys: ["TKT10002", "AYTKT10002", "581325", "A581325"],
-      },
-      {
-        name: "Ohjelmoinnin jatkokurssi",
-        keys: ["TKT10003", "AYTKT10003", "582103", "A582103"],
-      },
-      {
-        name: "Tietokantojen perusteet",
-        keys: ["TKT10004", "AYTKT10004", "581328", "A581328"],
-      },
-      {
-        name: "Ohjelmistotekniikan menetelmät",
-        keys: ["TKT20002", "AYTKT20002", "582104", "A582104"],
-      },
-    ],
-    aineopinnotPre2017: [
-      {
-        name: "Tietorakenteet ja algoritmit",
-        keys: ["TKT20001", "AYTKT20001", "58131", "A58131"],
-      },
-      {
-        name: "Käyttöjärjestelmät",
-        keys: ["TKT20003", "AYTKT20003", "582219", "A582219"],
-      },
-      {
-        name: "Tietokoneen toiminta",
-        keys: ["TKT10005", "AYTKT10005", "581305", "A581305"],
-      },
-      {
-        name: "Tietoliikenteen perusteet",
-        keys: ["TKT20004", "AYTKT20004", "582202", "A582202"],
-      },
-      {
-        name: "Laskennan mallit",
-        keys: ["TKT20005", "AYTKT20005", "582206", "A582206"],
-      },
-      {
-        name: "Ohjelmistotuotanto",
-        keys: ["TKT20006", "AYTKT20006", "581259", "A581259"],
-      },
-      {
-        name: "Ohjelmistotuotantoprojekti",
-        keys: ["TKT20007", "AYTKT20007", "581260", "A581260"],
-      },
-      {
-        name: "Kandidaatin tutkielma",
-        keys: ["TKT20013", "AYTKT20013", "582204", "A582204"],
-      },
-      {
-        name: "Kypsyysnäyte LuK",
-        keys: ["TKT20014", "AYTKT20014", "50036", "A50036"],
-      },
-    ],
-    labrat: [
-      {
-        name: "Aineopintojen harjoitustyö: Tietorakenteet ja algoritmit",
-        keys: ["TKT20010", "58161", "AYTKT20010", "A58161"],
-      },
-      {
-        name: "Aineopintojen harjoitustyö: Tietokantasovellus",
-        keys: ["TKT20011", "582203", "AYTKT20011", "A582203"],
-      },
-      {
-        name: "Aineopintojen harjoitustyö: Tietoliikenne",
-        keys: ["TKT20012", "AYTKT20012"],
-      },
-      {
-        name: "Aineopintojen harjoitustyö: Ohjelmointi",
-        keys: ["582221", "A582221"],
-      },
-    ],
-  },
-};
-
 const findFromKurssiTietokantaRecurse = ({ db, lyhenne }) =>
   Object.keys(db).reduce((acc, key) => {
     const courseFound =
@@ -241,7 +85,6 @@ const findFromKurssiTietokantaRecurse = ({ db, lyhenne }) =>
       db[key].find(({ keys }) =>
         keys.map(toLowerCase).includes(lyhenne.toLowerCase()),
       );
-
     return (
       acc ||
       (isArray(db[key])
@@ -786,7 +629,7 @@ const hommaaMeilleListaAsijoistaDommista = () => [
 const makeSomeStuff = (duplikaattiKurssit: string[]) =>
   hommaaMeilleListaAsijoistaDommista()
     .map((item) => [...Array.from(item.querySelectorAll("td"))])
-    .filter(elementArray => notEmptyList(elementArray)) // Filter empty lists
+    .filter((elementArray) => notEmptyList(elementArray)) // Filter empty lists
     .map((item) => map(item, "textContent").map(putsaaTeksti))
     .filter(([lyhenne]) => !duplikaattiKurssit.includes(lyhenne))
     .reverse()
@@ -802,10 +645,6 @@ const makeSomeStuff = (duplikaattiKurssit: string[]) =>
 // TODO: Typings
 const takeUntil = (list, n) =>
   list.reduce((initial, item, i) => (i < n ? [...initial, item] : initial), []);
-
-const sum = (a: number, b: number): number => a + b;
-
-const average = (list: number[]): number => list.reduce(sum, 0) / list.length;
 
 // TODO: Typings
 const annaMulleKeskiarvotKursseista = (stuff) =>
